backend/requirements.txt
fastapi>=0.115.0,<1.0.0
uvicorn[standard]>=0.32.0,<1.0.0
sqlalchemy[asyncio]>=2.0.36,<3.0.0
alembic>=1.14.0,<2.0.0
asyncpg>=0.30.0,<1.0.0
psycopg2-binary>=2.9.10,<3.0.0
aiomysql>=0.2.0,<1.0.0
openai>=1.55.0,<2.0.0
anthropic>=0.39.0,<1.0.0
sqlglot>=25.0.0,<27.0.0
PyJWT>=2.9.0,<3.0.0
redis>=5.2.0,<6.0.0
python-dotenv>=1.0.1,<2.0.0
bcrypt>=4.0.0,<6.0.0
cryptography>=44.0.0,<45.0.0
pydantic>=2.10.0,<3.0.0
pydantic-settings>=2.6.0,<3.0.0
httpx>=0.28.0,<1.0.0
PyYAML>=6.0.2,<7.0.0
